remove duplication log listener refactoring side effect code
