+++ com.hangum.tadpole.applicaion.start/launcher/tadpole_configuration.launch
@@ -16,7 +16,7 @@
 <booleanAttribute key="includeOptional" value="true"/>
 <booleanAttribute key="org.eclipse.jdt.launching.ATTR_USE_START_ON_FIRST_THREAD" value="true"/>
 <stringAttribute key="org.eclipse.jdt.launching.JRE_CONTAINER" value="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.7"/>
-<stringAttribute key="org.eclipse.jdt.launching.PROGRAM_ARGUMENTS" value="-os ${target.os} -ws ${target.ws} -arch ${target.arch}  -nl en -consoleLog -console  -registryMultiLanguage true -test -testDB -forceSystemInitialize -defaultDB MySQL -debuglog -passwd heechan.tadpole.owner.son &#13;&#10;-dbServer BZSmKIt+/Mwzm9tZHAgzrHenP+Nm9i0yK/RLZLvJ4v8rJWtjQjmR8Nx0+TE7r8XJtKlFHsbPnxhNiT+D86XWnsXtLGMJWLgb9F5Rrnuxrb2OKirsQM+/sydp2OfNK7d1M1W1zQd3d7OhAtnBmFjwcNd2otVizkgr&#10; -dbPath /Users/hangum/Documents/tadpole -OnlineServer"/>
+<stringAttribute key="org.eclipse.jdt.launching.PROGRAM_ARGUMENTS" value="-os ${target.os} -ws ${target.ws} -arch ${target.arch}  -nl en -consoleLog -console  -registryMultiLanguage true -test -testDB -forceSystemInitialize -defaultDB MySQL -debuglog -passwd startService.tdb.son heechan.tadpole.owner.son &#13;&#10;-dbServer x8KdY8QHvlYm5QLQify/K4GRIbZIynnk72dF2YcXj7puvlmtqGTb6mK2JjL01Q1Ku9imzbkaTwJ5PnaF+T2+RxN4PCj1/PXLIj9cLinjaa8kV3ViLIA8qnLKCQWenlJx3ZUcsFckR4lHGmAsO13yDMX3BwdmCX5YgRGInbsoOfOtu0OxJTdSoA==   Y3LvtpnhKIklOH7hD9AqJAClQwELIhEWryxRCASPicxfx2MjBprQ6iqzP6NbI4Y3pgZyOZWrdmYMbNH/cWqn5u0xFmDiNs+z+hHPHUfVoPkqywjtRxzIApmElJsel6/18/8/2jNdErMxM511lc6NwmBM5y8G4Ntgsydp2OfNK7d1M1W1zQd3d7OhAtnBmFjwcNd2otVizkgr&#10; -dbPath /Users/hangum/Documents/tadpole -OnlineServer"/>
 <stringAttribute key="org.eclipse.jdt.launching.SOURCE_PATH_PROVIDER" value="org.eclipse.pde.ui.workbenchClasspathProvider"/>
 <stringAttribute key="org.eclipse.jdt.launching.VM_ARGUMENTS" value="-Declipse.ignoreApp=true -Dosgi.noShutdown=true&#13;&#10;-Dorg.eclipse.equinox.http.jetty.log.stderr.threshold=info -Xverify:none -Xverify:none -XX:+UseParallelGC -XX:-UseConcMarkSweepGC -XX:PermSize=256M -XX:MaxPermSize=2014M -XX:MaxNewSize=4096M -XX:NewSize=128M -Xmx8292M -Dorg.eclipse.equinox.http.jetty.customizer.class=org.eclipse.rap.jettycustomizer.internal.SessionCookieCustomizer"/>
 <stringAttribute key="org.eclipse.rap.launch.browserMode" value="EXTERNAL"/>
