+++ com.hangum.tadpole.commons.libs/META-INF/MANIFEST.MF
@@ -29,6 +29,7 @@ Export-Package: .,
  com.hangum.tadpole.commons.libs.core.dao,
  com.hangum.tadpole.commons.libs.core.define,
  com.hangum.tadpole.commons.libs.core.googleauth,
+ com.hangum.tadpole.commons.libs.core.logs,
  com.hangum.tadpole.commons.libs.core.mails,
  com.hangum.tadpole.commons.libs.core.mails.dto,
  com.hangum.tadpole.commons.libs.core.mails.template,
