+++ com.hangum.tadpole.commons.libs/src/com/hangum/tadpole/commons/libs/core/logs/LogConfiguration.java
@@ -25,7 +25,7 @@ import org.eclipse.core.runtime.Platform;
  *
  */
 public class LogConfiguration {
-	String defaultFileName 	= "./logs/tadpole.log";
+	String defaultFileName 	= "../logs/tadpole/tadpole.log";
 	String filePattern 		= "%d %-5p [%t] %-17c{2} (%13F:%L) %3x - %m%n";
 	String defaultPattern 	= "%d{yyyy MMM dd HH:mm:ss, SSS} %p - %C{1} : %M %m %n";
