+++ com.hangum.tadpole.commons.libs/src/com/hangum/tadpole/commons/libs/core/utils/VelocityUtils.java
@@ -18,6 +18,7 @@ import java.util.Map;
 import org.apache.log4j.Logger;
 import org.apache.velocity.VelocityContext;
 import org.apache.velocity.app.Velocity;
+import org.apache.velocity.runtime.RuntimeConstants;
 
 /**
  * Velocity utils
@@ -40,6 +41,8 @@ public class VelocityUtils {
 		Writer writerReturn = new StringWriter();
 		
 		VelocityContext velocityCtx = new VelocityContext();
+		Velocity.setProperty(RuntimeConstants.RUNTIME_LOG, "../logs/tadpole/tadpoleVelocity.log");
+		
 		for(String strKey :mapParameter.keySet()) {
 			velocityCtx.put(strKey, mapParameter.get(strKey));
 		}
