+++ com.hangum.tadpole.commons/META-INF/MANIFEST.MF
@@ -31,7 +31,6 @@ Export-Package: com.google.gson,
  com.hangum.tadpole.commons.exception.dialog,
  com.hangum.tadpole.commons.google.analytics,
  com.hangum.tadpole.commons.help,
- com.hangum.tadpole.commons.log,
  com.hangum.tadpole.commons.start,
  com.hangum.tadpole.commons.util,
  com.hangum.tadpole.commons.util.download,
