+++ com.hangum.tadpole.commons/src/com/hangum/tadpole/commons/Activator.java
@@ -15,7 +15,7 @@ import org.eclipse.core.runtime.Platform;
 import org.osgi.framework.BundleActivator;
 import org.osgi.framework.BundleContext;
 
-import com.hangum.tadpole.commons.log.LogListener;
+import com.hangum.tadpole.commons.libs.core.logs.LogListener;
 
 public class Activator implements BundleActivator {
 	public static final String ID = "com.hangum.tadpole.commons.utils";
