+++ com.hangum.tadpole.rdb.core/src/com/hangum/tadpole/rdb/core/Activator.java
@@ -16,8 +16,8 @@ import org.eclipse.core.runtime.Platform;
 import org.eclipse.ui.plugin.AbstractUIPlugin;
 import org.osgi.framework.BundleContext;
 
-import com.hangum.tadpole.commons.log.LogConfiguration;
-import com.hangum.tadpole.commons.log.LogListener;
+import com.hangum.tadpole.commons.libs.core.logs.LogConfiguration;
+import com.hangum.tadpole.commons.libs.core.logs.LogListener;
 import com.hangum.tadpole.commons.start.TadpoleSystem;
 
 /**
