+++ com.hangum.tadpole.tajo.core/src/com/hangum/tadpole/tajo/core/Activator.java
@@ -5,8 +5,8 @@ import org.eclipse.core.runtime.Platform;
 import org.eclipse.ui.plugin.AbstractUIPlugin;
 import org.osgi.framework.BundleContext;
 
-import com.hangum.tadpole.commons.log.LogConfiguration;
-import com.hangum.tadpole.commons.log.LogListener;
+import com.hangum.tadpole.commons.libs.core.logs.LogConfiguration;
+import com.hangum.tadpole.commons.libs.core.logs.LogListener;
 
 /**
  * The activator class controls the plug-in life cycle
